+++ .eslintrc.json
@@ -19,6 +19,7 @@
     "import/no-named-as-default-member": "off",
     "jsx-a11y/no-static-element-interactions": "off",
     "jsx-a11y/click-events-have-key-events": "off",
+    "react/jsx-no-constructed-context-values": "off",
     "no-alert": "off"
   }
 }
